services/worker/pyproject.toml
[project]
name = "zebra-worker"
version = "0.0.0"
description = "FastAPI Worker Service for Zebra AI"
requires-python = ">=3.12"

dependencies = [
  "fastapi>=0.111.0",
  "uvicorn>=0.30.1",
  "rq>=1.16.2",
  "redis>=5.0.7",
  "python-dotenv>=1.0.1",
  "psycopg[binary]>=3.2.1",
  "psycopg-pool>=3.3.0",
]

[tool.uv]
managed = true
package = false
